Load fusionACS microdata dictionary

Returns a data frame containing metadata for the microdata variables in the local fusionACS database.

Usage

dictionary(directory = get_directory(), verbose = TRUE)

Arguments

directory

Path to the local fusionACS data directory.

Value

A data frame with the following columns:

variable

Variable name (column in microdata).

description

Variable description or definition.

survey

Acronym of the survey where the variable comes from.

vintage

Vintage of the donor survey where the variable comes from (NA for ACS variables; see 'years' instead).

respondent

Does the variable apply to "household" or "person" microdata observations?

type

The data type of the variable (e.g. integer).

n_values

Number of unique values the variable exhibits.

values

Summary of the values found in the microdata. For numeric variables, returns the output of summary. For categorical variables, returns a list of the unique values/levels.

years

Year(s) of ACS-PUMS microdata for which the variable is available.

custom

Logical indicating a 'custom' ACS variable created by the fusionACS team. See: https://github.com/ummel/fusionData/tree/master/survey-processed/ACS/custom

Examples

dict <- dictionary()
View(dict)