fusionACS: Integrated, extensive, and high-resolution microdata of U.S. households and individuals

This package provides access to data and tools to facilitate otherwise impossible U.S. social survey analyses. A documented pseudo-sample of the underlying fusionACS database is made available, along with functions to allow efficient microdata assembly and analysis. The fusionACS database is the result of statistical fusion of social survey microdata to simulate an extensive and high-resolution survey of U.S. households and individuals (Ummel et al. 2024, Tuccillo et al. 2023).

See also

Author

Maintainer: Kevin Ummel kevin.ummel@yale.edu

Other contributors:

Karthik Akkiraju karthik.akkiraju@yale.edu [contributor]