Get and set the number of cores

Control the number of cores used by fusionACS for parallel processing. The default (set at package load) is one less than the total number of cores detected by availableCores, which is generally accurate for both local machines and high-performance compute (HPC) clusters.

Usage

get_cores()

set_cores(n)

Arguments

n

Number of cores to use.

Value

Number of cores to use.